refactor: move file reading to main process
