+++ src/ipc-events.ts
@@ -31,4 +31,5 @@ export const enum IpcEvents {
   DELETE_SUGGESTION = 'DELETE_SUGGESTION',
   DELETE_SUGGESTIONS = 'DELETE_SUGGESTIONS',
   SUGGESTIONS_UPDATED = 'SUGGESTIONS_UPDATED',
+  READ_FILE = 'READ_FILE',
 }
