+++ src/renderer/components/log-content.tsx
@@ -9,7 +9,7 @@ import React from 'react';
 import { LogLineDetails } from './log-line-details/details';
 import { Scrubber } from './scrubber';
 import { getFontForCSS } from './preferences-font';
-import { getTypeForFile } from '../processor';
+import { getTypeForFile } from '../../utils/get-file-types';
 import { NetLogView } from './net-log-view';
 import { ToolView } from './tool-view';
 import { LogTimeView } from './log-time-view';
